Fixed linting and moved logging setup to own file
