+++ golang/cmd/club/logging.go
@@ -1,0 +1,30 @@
+package main
+
+import (
+	"fmt"
+	"path"
+	"runtime"
+	"strings"
+
+	"github.com/sirupsen/logrus"
+)
+
+func initLogging(config *Config) {
+	if config.Environment == EnvironmentProduction {
+		logrus.SetLevel(logrus.InfoLevel)
+	} else {
+		logrus.SetLevel(logrus.DebugLevel)
+	}
+
+	// Setup logging
+	logrus.SetReportCaller(true)
+	logrus.SetFormatter(&logrus.JSONFormatter{
+		CallerPrettyfier: func(f *runtime.Frame) (string, string) {
+			s := strings.Split(f.Function, ".")
+			funcName := s[len(s)-1]
+			return funcName, fmt.Sprintf("%s:%d", path.Base(f.File), f.Line)
+		},
+	})
+
+	logrus.Infof("Log level: %s", logrus.GetLevel())
+}
